+++ packages/openapi-ts/test/__snapshots__/2.0.x/security-api-key/sdk.gen.ts
@@ -7,7 +7,6 @@ export const client = createClient(createConfig());
 
 export const getFoo = <ThrowOnError extends boolean = false>(options?: Options<GetFooData, ThrowOnError>) => {
     return (options?.client ?? client).get<unknown, unknown, ThrowOnError>({
-        ...options,
         security: [
             {
                 in: 'query',
@@ -15,6 +14,7 @@ export const getFoo = <ThrowOnError extends boolean = false>(options?: Options<G
                 type: 'apiKey'
             }
         ],
-        url: '/foo'
+        url: '/foo',
+        ...options
     });
 };