+++ packages/openapi-ts/test/__snapshots__/2.0.x/security-basic/sdk.gen.ts
@@ -7,13 +7,13 @@ export const client = createClient(createConfig());
 
 export const getFoo = <ThrowOnError extends boolean = false>(options?: Options<GetFooData, ThrowOnError>) => {
     return (options?.client ?? client).get<unknown, unknown, ThrowOnError>({
-        ...options,
         security: [
             {
                 scheme: 'basic',
                 type: 'http'
             }
         ],
-        url: '/foo'
+        url: '/foo',
+        ...options
     });
 };