+++ examples/openapi-ts-fastify/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @example/openapi-ts-fastify
 
+## 0.1.0
+
+### Minor Changes
+
+- [#1568](https://github.com/hey-api/openapi-ts/pull/1568) [`465410c`](https://github.com/hey-api/openapi-ts/commit/465410c201eb19e737e3143ad53a146e95f80107) Thanks [@mrlubos](https://github.com/mrlubos)! - feat: change the default parser
+
+### Patch Changes
+
+- Updated dependencies [[`465410c`](https://github.com/hey-api/openapi-ts/commit/465410c201eb19e737e3143ad53a146e95f80107)]:
+  - @hey-api/client-fetch@0.7.0
+
 ## 0.0.11
 
 ### Patch Changes